refactor: split server.ts into src/mcp/, src/http/, src/server/ modules
Split 4,167-line src/server.ts into focused modules:
- src/server/types.ts: ServerOptions, ServerDeps, ResolvedWorkspace interfaces
- src/server/utils.ts: resolveWorkspace, formatters, shared helpers
- src/mcp/tools/types.ts: McpToolContext interface for closure injection
- src/mcp/tools/memory.ts: 16 memory MCP tools
- src/mcp/tools/graph.ts: 7 graph MCP tools
- src/mcp/tools/code.ts: 6 code intelligence MCP tools
- src/mcp/tools/indexing.ts: 1 indexing MCP tool
- src/mcp/index.ts: createMcpServer() composed from register* functions
- src/http/sse.ts: SSE session management
- src/http/routes.ts: HTTP pathname dispatch (handleRequest)
- src/http/server.ts: createHttpServer()
- src/server/bootstrap.ts: startServer(), createRejectionThreshold()
- src/server.ts: 4-line barrel shim

Also adds OpenSpec proposal/design/specs/tasks for Phase 3.
Updates test/rri-t-round4.test.ts SSE path to src/http/sse.ts.

Zero behavior changes. Tests: 1518 passing (10 pre-existing watcher failures).
