+++ openspec/changes/nano-brain-phase3-server-split/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-05
